The Entire Project's Front End
Though I completed this project with a Biology student, he had no concept of CS or programming and basically shrugged at me as I did the entire thing. The Front End is HTML, CSS, and Bootstrap. The Back End services were all done in PERL, which query the database. There were additional PERL files that only ran once -- they parsed the massive genome and populated a MySQL database. This was basically a LAMP stack application, but the server it once resided on is now defunct. You can't really use it as intended, but you could pull the Front End files at least and view them in a browser.